Add k8s deployment ymls for bazelremote (#9)
PoC of bazel remote service setup for linux rocm runners in `Saiscale` cluster.
Still needs to be productionized, and moved to windows.

* Add k8s deployment ymls for bazelremote

Ran these commands to setup on SaiScale k8s cluster
```
kubectl create namespace bazelremote-ns
kubectl apply -f azure-linux-scale-rocm-bazelremote.yml -n bazelremote-ns
kubectl apply -f azure-linux-scale-rocm-bazelremote-svc.yml -n bazelremote-ns
```

* Use `app` label from deployed k8s yml for svc yml

there was a mismatch in the yml of the k8s bazelremote-svc, which was looking for the label `app.kubernetes.io/name` but it was set with the `app` label in `azure-linux-scale-rocm-bazelremote.yml`

* Disable verification to make ccache compatible

* Note that max size is in GiB and set to 1TiB

* Make env value a string in k8s deployment yml

 kubectl apply -f azure-linux-scale-rocm-bazelremote.yml -n bazelremote-ns

* pin bazelremote node w/ nodeselector, tolerations

* Enable bazel remote metrics (applied on 20250703)